# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "center API",
    "description": "Coder by Hong Dang"
  },
  "host": "localhost:3001",
  "basePath": "/api/v1.0/",
  "tags": [
    {
      "name": "nodejs example",
      "description": "nodejs example"
    }
  ],
  "schemes": [],
  "securityDefinitions": {
    "Bearer": {
      "name": "Authorization",
      "in": "header",
      "type": "apiKey"
    },
    "Language": {
      "name": "LanguageCode",
      "in": "header",
      "type": "apiKey"
    }
  },
  "consumes": [],
  "produces": [],
  "paths": {
    "/user/register": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to register",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "User name + password",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/authenticaion"
            }
          }
        ],
        "responses": {},
        "security": [
          {
            "Bearer": [],
            "LanguageCode": []
          }
        ]
      }
    },
    "/user/login": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to get token with username & password",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "User name + password",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/login"
            }
          }
        ],
        "responses": {},
        "security": [
          {
            "Bearer": [],
            "LanguageCode": []
          }
        ]
      }
    },
    "/user/getListUser": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to get list user do not use pagging",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "Bearer": [],
            "LanguageCode": []
          }
        ]
      }
    },
    "/user/getListUserOnPage": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to get list of user using pagging",
        "parameters": [
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "currentPage",
            "in": "query",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {},
        "security": [
          {
            "Bearer": [],
            "LanguageCode": []
          }
        ]
      }
    },
    "/user/getMyInfo": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to get my info",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "Bearer": [],
            "LanguageCode": []
          }
        ]
      }
    },
    "/user/updateMyInfo": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to update my info: provide:oldName, oldEmail",
        "parameters": [
          {
            "name": "new information",
            "in": "body",
            "description": "New information ",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/updateMyInfo"
            }
          }
        ],
        "responses": {},
        "security": [
          {
            "Bearer": [],
            "LanguageCode": []
          }
        ]
      }
    }
  },
  "definitions": {
    "Parents": {
      "type": "object",
      "properties": {
        "father": {
          "type": "string",
          "example": "Simon Doe"
        },
        "mother": {
          "type": "string",
          "example": "Marie Doe"
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Jhon Doe"
        },
        "age": {
          "type": "number",
          "example": 29
        },
        "parents": {
          "$ref": "#/definitions/Parents"
        },
        "diplomas": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "school": {
                "type": "string",
                "example": "XYZ University"
              },
              "year": {
                "type": "number",
                "example": 2020
              },
              "completed": {
                "type": "boolean",
                "example": true
              },
              "internship": {
                "type": "object",
                "properties": {
                  "hours": {
                    "type": "number",
                    "example": 290
                  },
                  "location": {
                    "type": "string",
                    "example": "XYZ Company"
                  }
                }
              }
            }
          }
        }
      }
    },
    "authenticaion": {
      "type": "object",
      "properties": {
        "Name": {
          "type": "string",
          "example": "Nguyen Van D"
        },
        "Email": {
          "type": "string",
          "example": "nvd@gmail.com"
        },
        "Password": {
          "type": "string",
          "example": "123456"
        },
        "RePassword": {
          "type": "string",
          "example": "123456"
        }
      },
      "required": [
        "Name",
        "Email",
        "Password",
        "RePassword"
      ]
    },
    "userActive": {
      "type": "object",
      "properties": {
        "userID": {
          "type": "string",
          "example": "611c53f5-84c0-4911-a511-ef0279a4bbd9"
        }
      },
      "required": [
        "userID"
      ]
    },
    "forgetPassword": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "nvd246@gmail.com"
        }
      },
      "required": [
        "email"
      ]
    },
    "login": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "nguyenvana@gmail.com"
        },
        "password": {
          "type": "string",
          "example": "123456"
        }
      },
      "required": [
        "email",
        "password"
      ]
    },
    "createNewPassword": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string",
          "example": "6AHfsQbHA39u9BIe4Jti"
        }
      },
      "required": [
        "code"
      ]
    },
    "updateMyInfo": {
      "type": "object",
      "properties": {
        "OldEmail": {
          "type": "string",
          "example": "nguyenvana@gmail.com"
        },
        "OldName": {
          "type": "string",
          "example": "nguyen van a"
        },
        "NewEmail": {
          "type": "string",
          "example": "nguyenvana@gmail.com"
        },
        "NewName": {
          "type": "string",
          "example": "nguyen van a"
        },
        "NewPassword": {
          "type": "string",
          "example": "123456"
        },
        "NewRePassword": {
          "type": "string",
          "example": "123456"
        }
      },
      "required": [
        "OldEmail",
        "OldName",
        "NewEmail",
        "NewName",
        "NewPassword",
        "NewRePassword"
      ]
    }
  }
}